Version_1.py
With this code, the bot is completely reliant on the camera. It is able to avoid the walls and make the major turns. The bot glitches time to time with the extended border. With this code the bot does not start with a button press nor does it stop after 3 laps.